+++ grand_trade_auto/apic/__init__.py
@@ -1,5 +1,6 @@
 # pylint: disable=missing-module-docstring
 __all__ = [
-        'test_broker_alpaca',
-        'test_brokers',
+        'alpaca',
+        'apic_meta',
+        'apics',
 ]